modules/function/img_processing/requirements.txt
google-cloud-storage==2.13.0
google-cloud-pubsub==2.18.4
Pillow==10.1.0
grpcio==1.59.3
pymongo==4.6.0
Flask==2.3.3